pyproject.toml
[project]
name = "artemis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.26.0",
    "pyntcore>=2026.2.2",
]
